requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
supabase==2.7.4
openai==1.46.1
pdfplumber==0.11.4
python-multipart==0.0.9
